Change minTimeInc (for calibration sampling) into an int and set it to zero (we don't need to pause taking samples)

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/base/calibration/MotorCalibConfig.java
@@ -34,7 +34,7 @@ import androidx.annotation.NonNull;
 import org.firstinspires.ftc.teamcode.base.config.Validatable;
 
 public class MotorCalibConfig implements Validatable {
-    public double  minTimeInc;
+    public int     minTimeInc;
     public int     timeResolution;
     public int     powerResolution;
     public double  averagingTime;
